% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_factors_to_char.R
\name{all_factors_to_char}
\alias{all_factors_to_char}
\title{Convert all factors in a data.frame to character}
\usage{
all_factors_to_char(dat)
}
\arguments{
\item{dat}{A data.frame}
}
\value{
The input data.frame with all factors converted to character
}
\description{
Converts all factors in a data.frame to character.
}
\examples{
sapply(iris, class)
iris2 <- all_factors_to_char(iris)
sapply(iris2, class)

}
