% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_factors_to_num.R
\name{all_factors_to_num}
\alias{all_factors_to_num}
\title{Convert all factors in a data.frame to numeric}
\usage{
all_factors_to_num(dat)
}
\arguments{
\item{dat}{A data.frame}
}
\value{
The input data.frame with all factors converted to numeric
}
\description{
Converts all factors in a data.frame to numeric.
This function is to be used when all factors have levels representing numeric values.
Otherwise, the function returns missings.
}
\examples{
sapply(iris, class)
iris2 <- all_factors_to_char(iris)
sapply(iris2, class)

}
