% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Create a bootstrap sample or return a bootstrapped statistic}
\usage{
bootstrap(dat, n = nrow(dat), rep = 1, fun = NULL)
}
\arguments{
\item{dat}{A data.frame or matrix.}

\item{n}{The sample size of the bootstrap sample. Same size as the original sample by default.}

\item{rep}{The number of times a bootstrap sample should be taken. If >1, returns a list.}

\item{fun}{An expression as a character string, to be evaluated internally. The data should be assumed attached.}
}
\value{
A bootstrapped data.frame/matrix, or a numeric vector with a bootstrapped statistic.
}
\description{
Create a bootstrap sample or return a bootstrapped statistic
}
\examples{
# bootstrapped coefficient from a regression model
bootstrap(iris, rep = 100, fun = 'coef(lm(Sepal.Length ~ Species))[2]')

# bootstrapped mean
bootstrap(iris, rep = 100, fun = 'mean(Petal.Length)')

# bootstrap sample
iris_boot <- bootstrap(iris)
}
