% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_endpoints.R
\name{combine_endpoints}
\alias{combine_endpoints}
\title{Combine two events and times for time-to-event data}
\usage{
combine_endpoints(dat, events, times, add = FALSE, add.names = NULL)
}
\arguments{
\item{dat}{A data.frame or matrix.}

\item{events}{The event columns (a numeric/character vector of length 2) to combine}

\item{times}{The time columns (a numeric/character vector of length 2) to combine.}

\item{add}{If TRUE, adds the truncated time and recoded event variables to the input data.frame.}

\item{add.names}{A character vector with the colnames for the columns added if add=TRUE.}
}
\description{
Combines two events with times for time-to-event data.
}
