% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_quantiles.R
\name{cut_quantiles}
\alias{cut_quantiles}
\title{Cut a numeric vector by specified percentiles}
\usage{
cut_quantiles(var, percentiles, labels = NULL)
}
\arguments{
\item{var}{A numeric vector.}

\item{percentiles}{A vector of percentiles to cut by.}

\item{labels}{If not NULL, these are the levels for the resulting factor.}
}
\value{
A factor representing the categories after cutting the vector.
}
\description{
Cuts a numeric vector by specified percentiles.
}
\examples{
cut_quantiles(iris$Sepal.Length, c(0,0.33,0.66,1),c('Low','Middle','High'))

}
