% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Exclude observations from a sample based on logical conditions}
\usage{
exclude(dat, condition, verbose = TRUE)
}
\arguments{
\item{dat}{A data.frame or matrix.}

\item{condition}{A logical condition specifying the exclusion criteria.}

\item{verbose}{If TRUE, prints how many rows were dropped based on the exclusion criteria.}
}
\value{
A subset of the original data.frame or matrix with the exclusion criteria applied.
}
\description{
Does the opposite of subset().
}
\examples{
exclude(iris, Species == 'setosa' | Sepal.Length > 6)
}
