% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersperse.R
\name{intersperse}
\alias{intersperse}
\title{Interleave two data.frames or matrices row by row}
\usage{
intersperse(df1, df2, add_df_id = TRUE)
}
\arguments{
\item{df1}{Input data.frame or matrix.}

\item{df2}{A second data.frame or matrix of equal dimensions as df1.}

\item{add_df_id}{If TRUE, adds an additional column indicating the original data.frame.}
}
\value{
A combined data.frame.
}
\description{
Interleave two data.frames or matrices of equal dimensions, row by row.
}
\examples{
iris_s1 <- iris[1:5, 1:4]
iris_s2 <- iris[1:5, 1:4] - 0.5
intersperse(iris_s1, iris_s2)

}
