% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{kmers}
\alias{kmers}
\title{Generate k-mers from a string}
\usage{
kmers(x, k = 3)
}
\arguments{
\item{x}{A character string.}
}
\value{
Returns a character vector of length u^k, where u is the number of unique characters appearing across k-mers.
}
\description{
Retrieves k-mers from a character string.
}
\examples{
kmers('abcd', kmers = 4)
}
