% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_do.R
\name{matrix_do}
\alias{matrix_do}
\title{Perform a function in a nested for loop}
\usage{
matrix_do(outer = NULL, inner = NULL, fun = NULL, progress = TRUE)
}
\arguments{
\item{outer}{Argument denoting what to iterate over for the outer loop.}

\item{inner}{Argument denoting what to iterate over for the inner loop.}

\item{fun}{An anonymous function with 2 arguments.}

\item{progress}{If TRUE, prints the iteration of all iterations outer * inner.}
}
\value{
A list containing the output of the nested for loop.
}
\description{
Perform a function within a nested for loop.
}
\examples{
matrix_do(outer=c(1,2,3),inner=letters[1:5], fun=function(x,y) print(paste0(x,y)))

}
