% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_column.R
\name{move_column}
\alias{move_column}
\title{Move a column to a new position}
\usage{
move_column(dat, column, to)
}
\arguments{
\item{dat}{Input data.frame or matrix.}

\item{column}{The name or index of the column to be moved.}

\item{to}{The column index to which the column should be moved.}
}
\value{
A data.frame with the column in the new position.
}
\description{
Moves a column to a designated position.
}
\examples{
head(iris)
head(move_column(iris, 5, 1))
head(move_column(iris, 'Petal.Length', 5))

}
