% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_out.R
\name{partial_out}
\alias{partial_out}
\title{Decorrelate a variable by one or several others using OLS}
\usage{
partial_out(dat, var, decor_vars)
}
\arguments{
\item{dat}{A data.frame or matrix.}

\item{var}{The variable (a character vector of length 1) to be decorrelated.}

\item{decor_vars}{The variable(s) (character vector) to decorrelate by.}
}
\value{
The decorrelated variable.
}
\description{
Partials out variables by linear regression.
}
\examples{
cor.test(iris$Petal.Length, iris$Sepal.Length)
decor_sepl <- partial_out(iris, 'Sepal.Length', 'Petal.Length')
cor.test(iris$Petal.Length, decor_sepl)

}
