% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_diff.R
\name{predict_diff}
\alias{predict_diff}
\title{Predict difference in outcome for a pair of linear predictors}
\usage{
predict_diff(
  mod,
  x1,
  x2,
  type = NULL,
  ratio = FALSE,
  ci = FALSE,
  data = NULL,
  boot.iter = 1000,
  decimals = 3,
  plot = TRUE,
  seed = NULL
)
}
\arguments{
\item{mod}{The regression model based on which to generate the predictions.}

\item{x1}{One row in X to compare against another row in X.}

\item{x2}{One row in X to compare against another row in X}

\item{type}{The type of prediction. By default the difference is on the scale of the linear predictor.}

\item{ratio}{If TRUE, returns the ratio of predictions instead of the difference.}

\item{ci}{If TRUE, returns a bootstrapped confidence interval (95%).}

\item{data}{If ci is TRUE, the data based on which the bootstrap samples are to be made (same as the data that informed the model) should be provided.}

\item{boot.iter}{Number of bootstrap samples.}

\item{decimals}{Number of decimals to display for the difference and its confidence interval.}

\item{plot}{If TRUE, plots the distribution of bootstrap differences based on which the confidence interval is created.}

\item{seed}{If not NULL, sets a random seed before bootstrapping.}
}
\description{
Gives the difference in predicted y for two rows in X (X = x1 vs X = x2), based on a given regression model. Returns a confidence interval for the difference based on bootstrap samples. A ratio of predictions can also be returned.
}
\examples{
predict_diff(lm(Sepal.Length ~., iris), x1 = iris[1,-1], x2 = iris[4,-1], ci = TRUE, data = iris)

# what is the predicted difference in Sepal.Length between different species?
x0 <- iris[1,-1] # Species is setosa
x1 <- x0
x1$Species <- 'virginica'
x2 <- x0
x2$Species <- 'versicolor'
predict_diff(lm(Sepal.Length ~., iris), x1 = x1, x2 = x0, ci = TRUE, data=iris)
predict_diff(lm(Sepal.Length ~., iris), x1 = x1, x2 = x2, ci = TRUE, data=iris)
predict_diff(lm(Sepal.Length ~., iris), x1 = x2, x2 = x0, ci = TRUE, data=iris)
predict_diff(lm(Sepal.Length ~., iris), x1 = x2, x2 = x1, ci = TRUE, data=iris)

}
