% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartiles.R
\name{quartiles}
\alias{quartiles}
\title{Cut a numeric vector by quartiles of the vector.}
\usage{
quartiles(var)
}
\arguments{
\item{var}{A numeric vector.}

\item{labels}{If not NULL, these are the levels for the resulting factor.}
}
\value{
A factor representing the categories after cutting the vector.
}
\description{
Cuts a numeric vector by quartiles of the vector.
}
\examples{
quartiles(iris$Sepal.Length,c('Low','Middle low','Middle high','High'))

}
