% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_factor.R
\name{recode_factor}
\alias{recode_factor}
\title{Recode the levels of a factor}
\usage{
recode_factor(var, new_levels = levels(var))
}
\arguments{
\item{var}{A factor.}

\item{new_levels}{The new levels of the factor, in order. The number of levels should be
equivalent to that of the original factor, but levels need not be unique.}
}
\value{
The recoded factor.
}
\description{
Recodes a factor by relabeling it.
}
\examples{
table(iris$Species)
table(recode_factor(iris$Species, new_levels=c('Setosa','Not setosa','Not setosa')))

}
