% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename one or several columns}
\usage{
rename_columns(dat, column, to)
}
\arguments{
\item{dat}{Input data.frame or matrix.}

\item{column}{The name(s) of the column(s) to be renamed}

\item{to}{The new name(s).}
}
\value{
A data.frame with the renamed column(s)
}
\description{
Renames one or multiple columns in a data.frame or matrix.
}
\examples{
head(iris)
head(rename_column(iris, 'Species', 'Renamed'))

}
