% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Calculate skewness of a vector}
\usage{
skew(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Returns the skewness of the input vector.
}
\description{
Calculates the skewness of a continuous variable
}
\examples{
skew(iris$Sepal.Length)
}
