% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_df.R
\name{sort_df}
\alias{sort_df}
\title{Reorder a data.frame/matrix by a column.}
\usage{
sort_df(dat, sort_by = NULL, decreasing = TRUE)
}
\arguments{
\item{dat}{A data.frame or matrix.}

\item{sort_by}{The column index or name to sort the data.frame by.}

\item{decreasing}{If TRUE, sorts in decreasing order.}
}
\value{
The reordered data.frame or matrix.
}
\description{
Reorders a data.frame or matrix by a specified column.
}
\examples{
head(iris)
head(sort_df(iris, 'Petal.Length'))

}
