% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_column.R
\name{swap_column}
\alias{swap_column}
\title{Swap two columns in a data.frame or matrix}
\usage{
swap_column(dat, column, to)
}
\arguments{
\item{dat}{A data.frame or matrix.}

\item{column}{A column, by name or by index, to be swapped in position with another.}

\item{to}{Index for the second column, which will be swapped with the column named in 'column'.}
}
\value{
A data.frame with the indicated columns swapped.
}
\description{
Swaps two columns in a data.frame or matrix.
}
\examples{
head(iris)
head(swap_column(iris, 5, 1))
head(swap_column(iris, 'Petal.Length', 2))

}
