% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tertiles.R
\name{tertiles}
\alias{tertiles}
\title{Cut a numeric vector by tertiles of the vector.}
\usage{
tertiles(var, labels = NULL)
}
\arguments{
\item{var}{A numeric vector.}

\item{labels}{If not NULL, these are the levels for the resulting factor.}
}
\value{
A factor representing the categories after cutting the vector.
}
\description{
Cuts a numeric vector by tertiles of the vector.
}
\examples{
tertiles(iris$Sepal.Length,c('Low','Middle','High'))

}
