% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_normal.R
\name{transform_to_normal}
\alias{transform_to_normal}
\title{Transform a continuous variable to approximately normal distribution}
\usage{
transform_to_normal(x, constant = 1, rank = FALSE)
}
\arguments{
\item{x}{A vector.}
}
\value{
Returns the transformed vector.
}
\description{
Tranforms a continuous variable to be roughly normally distributed, picking the transformation that results in the lowest absolute skewness.
}
\examples{
hist(iris$Sepal.Length)
hist(transform_to_normal(iris$Sepal.Length))
hist(transform_to_normal(iris$Sepal.Length, rank=TRUE))
}
