% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_time.R
\name{truncate_time}
\alias{truncate_time}
\title{Truncate survival times}
\usage{
truncate_time(
  dat,
  event,
  time,
  max_follow_up = 1,
  add = FALSE,
  add.names = NULL
)
}
\arguments{
\item{dat}{A data.frame or matrix.}

\item{event}{The event (a numeric/character vector of length 1) to recode.}

\item{time}{The time (a numeric/character vector of length 1) to truncate.}

\item{max_follow_up}{The follow-up time to truncate to.}

\item{add}{If TRUE, adds the truncated time and recoded event variable to the input data.frame.}

\item{add.names}{A character vector with the colnames for the columns added if add=TRUE.}
}
\description{
Truncates survival times and recodes the corresponding event.
}
