% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorizing outliers}
\usage{
winsorize(dat, mads = 4)
}
\arguments{
\item{dat}{A numeric vector.}

\item{mads}{The number of MADs away from the median defining a univariate outlier.}
}
\value{
The winsorized vector.
}
\description{
Winsorizes univariate outliers, defined by a specified number of median absolute deviations from the median.
}
\examples{
set.seed(123)
x <- rnorm(1000)
x <- c(x, 5.1, 6.2, 8.9, 10.1, 12.2, -14, -18.4, -7.2) # add outliers
hist(x)
hist(winsorize(x)) # after winsorizing outliers

}
